pyproject.toml
[project]
name = "oumaji-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.4",
    "mlflow>=2.22.0",
    "pandas>=2.2.3",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "pyyaml>=6.0.2",
]
